requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
inflection==0.5.1
jsonschema==4.17.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyMySQL==1.0.2
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
whitenoise==6.2.0
